refactor(core): address token-engine instrumentation review
- Split static ramps (STATUS_RAMP / CHART_* / NEUTRAL) out of the
  surface-named module into static-ramps.ts
- Move Mode / NexusSurfaceTone to palette.ts so surface-ladder is a true
  leaf; drop the ramp re-export shim in derive-theme and repoint imports
- Drop the unused, double-counting contrast param from anchorToStep
- Remove dead ShadeAnchor / SurfaceSteps public exports and the
  commented-out SURFACE_STEP_OVERRIDES block
- Drop the dead PERCEPTUAL_L_GRID spread so unknown shades fail fast
- Remove the tautological surface-ladder step assertions
- Extract shared engine-snapshot-matrix.mjs so the fixture generator and
  test no longer duplicate (and drift) the shaping helpers
