+++ packages/core/src/lib/palette.ts
@@ -38,6 +38,12 @@ export const TIER_THRESHOLDS = {
 
 export type Tier = keyof typeof TIER_THRESHOLDS;
 
+/** Light/dark theme axis. */
+export type Mode = 'light' | 'dark';
+
+/** Surface tone families a Nexus theme can be seeded with. */
+export type NexusSurfaceTone = 'stone' | 'neutral' | 'zinc' | 'slate' | 'gray';
+
 export const PERCEPTUAL_L_GRID: Record<Shade, number> = Object.freeze(
   perceptualGrid as Record<Shade, number>
 );
